Adding two new projects. libCore, which will be used by all future projects USBCoreBlink, which is what's left USBBlink project after the core stuff is taken out.
Still needs separation. main() is in the project not the lib, which i dont like.
So is the USBJumpTable.asm. It really should be in libCore, but it doesnt link if it's put there. I may need to port it to C (if i can) to get it to work.